implement table/recording POST, GET, DELETE (local xlsx for storage)
